backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
gspread==5.9.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
oauth2client==4.1.3
python-dateutil==2.8.2
